makefile: Fix BIBSRC definition
The glob pattern passed to find must be enclosed within quotes.
In the abscence of quotes, *.bib matches bibliography.bib when
this file is on current directory. So, instead of "*.bib", find
sees "bibliography.bib", and searches for files with this name
inside bib/ subdirectory instead of finding all files matching
*.bib.

This prevents the update of bibliography.bib when some file is
modified inside bib/ subdirectory, since BIBSRC will be empty.

Also, the assignment should be with := instead of = to prevent
multiple executions of find.