Make the makefile actively register its targets
Previously, every project that included this makefile had to register
this makefile's targets as dependencies of their own targets
(for instance, all had to depend on bib-all).
This update makes the makefile do this registering.

We renamed the makefile to makefile.mk to make this intention
more explicit. However, this renaming forces us to depend
on other makefiles. To retain "standalone-ness", a file named "makefile"
was added to this repository, that simply include makefile.mk.