Fixed encoding in the url field
Some urls have characters that are special to LaTeX, like '~' and '_';
I used '\_' to encode '_' and '{\textasciitilde}' to encode '~'.
I did this encoding to use the ABNT style,
which exposes the URL field "naked" to LaTeX.
Thus, '_', which is only allowed in math mode,
and '~', which is the non-breaking space,
had to be quoted.

Ideally, the url field should be quoted automatically,
using either the url package, or the hyperref package.
ABNTex seems to support hyperref,
if hyperref is included before abntcite.

But this "solves the problem too much": now even the characters '\_'
and '{\textasciitilde}' are quoted. For instance, Blum1967's entry reads
    http://port70.net/\textasciitildensz/articles/classic/blum\_co
(for some reason, the text gets truncated).
Thus, in order to use hyperref with abntcite,
'~' and '_' must be "encoded" directly as '~' and '_'.

Since most of the bibliography styles that uses the url field
enclose it in a '\url' command (from url package),
there will be no harm in letting this code unescaped
in the .bib files.